refactor(wasm): split zero_copy.rs into mesh + symbolic DTO modules
zero_copy.rs (1318 LOC) held two independent #[wasm_bindgen] DTO families.
Convert it into a zero_copy/ directory: mesh.rs (MeshDataJs / MeshCollection
+ the Z-up->Y-up conversion) and symbolic.rs (Symbolic* + collection), with
mod.rs declaring the submodules, keeping the free `get_memory` binding, and
re-exporting every type.

Pure code motion with zero visibility changes (each field reconstruction stays
with its struct). All #[wasm_bindgen] attributes/js_names/getters preserved, so
the generated bindings are unaffected. crate::zero_copy::* and the lib.rs
re-export resolve unchanged; lib.rs untouched. Identical 9-warning baseline,
no new warnings; cargo test -p ifc-lite-wasm green (12); wasm32 build clean.
